requirements.txt
numpy==1.22
pandas==2.0.1
tensorflow==2.12.0
scikit-learn==1.2.2
opencv-python==4.7.0.72
absl-py==1.4.0